backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pycryptodome==3.22.0
pytest==8.3.5
python-dotenv==1.1.0
Werkzeug==3.1.3
